Refactor loan package to transaction package for naming consistency
