+++ src/main/java/seedu/address/model/util/SampleDataUtil.java
@@ -7,8 +7,8 @@ import java.util.stream.Collectors;
 
 import seedu.address.model.AddressBook;
 import seedu.address.model.ReadOnlyAddressBook;
-import seedu.address.model.loan.Loan;
-import seedu.address.model.loan.MonthlyLoan;
+import seedu.address.model.transaction.Loan;
+import seedu.address.model.transaction.MonthlyLoan;
 import seedu.address.model.person.Address;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.Name;
