+++ src/test/java/seedu/address/model/transaction/MonthlyLoanTest.java
@@ -1,4 +1,4 @@
-package seedu.address.model.loan;
+package seedu.address.model.transaction;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
